requirements.txt
openai>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
sentence-transformers>=2.7.0
matplotlib>=3.7.0
seaborn>=0.13.0
joblib>=1.3.0
pyyaml>=6.0
jupyter>=1.0.0
ipykernel>=6.25.0
